Allow same tag for "--sjdbGTFtagExonParentTranscript" and "--sjdbGTFtagExonParentGene".
Allow to specify same tag for "--sjdbGTFtagExonParentTranscript" and
"--sjdbGTFtagExonParentGene" when reading a GTF file.

Before this patch, if both options had the same tag.

In case of STAR solo:
  - "*.Solo.out/genes.tsv" would be empty
  - "*.Solo.out/matrix.mtx" assigns all counted reads to the first
    "gene" instead of the correct one.

Supporting this feature makes is a bit easier to use STAR to use
regions as genes:

    chr1    regions_test    Region  1000       2000        .       .       .       region "chr1:1000-2000";
    chr2    regions_test    Region  3000       4000        .       .       .       region "chr2:3000-4000";

With the old code the fake GTF file needed to look like this:

    chr1    regions_test    Region  1000       2000        .       .       .       region_gene "chr1:1000-2000"; region_transcript "chr1:1000-2000";
    chr2    regions_test    Region  3000       4000        .       .       .       region_gene "chr2:3000-4000"; region_transcript "chr2:3000-4000";