Fix memory leaks in ReadAlign.
Use unique_ptr to manage memory in ReadAlign and orbit front-end.  Try
to do this in a minimally-invasive way so we don't have to touch too
much other code.

Take a few things that shouldn't have been pointers and make them
not be pointers.

Slab-allocate the storage for some arrays-of-arrays.

Make argv parameters const, for both the c++ side and the rust side.
Clean up memory management for C strings on the rust side.